fix(FEC-10968): OTT doesn't have default external-stream-redirect-helper
